Port to latest nbgrader (>=0.5.0)
Starting in version 0.5.0 of nbgrader, the formgrader is no longer a
standalone command. See
http://nbgrader.readthedocs.io/en/stable/configuration/jupyterhub_config.html
for details.

NOTE: The way we are deploying nbgrader_config.py in this commit is not
ideal. Currently, we are not able to support more than one course per
grader. Again, see the link above for details.